+++ src/ace_crc/crc8_bit.cpp
@@ -1,0 +1,55 @@
+/**
+ * \file
+ * Functions and types for CRC checks.
+ *
+ * Generated on Thu Dec  3 21:05:47 2020
+ * by pycrc v0.9.2, https://pycrc.org
+ * using the configuration:
+ *  - Width         = 8
+ *  - Poly          = 0x07
+ *  - XorIn         = 0x00
+ *  - ReflectIn     = False
+ *  - XorOut        = 0x00
+ *  - ReflectOut    = False
+ *  - Algorithm     = bit-by-bit-fast
+ *
+ * Auto converted to Arduino C++ on Thu Dec  3 21:05:47 PST 2020
+ * by AceCRC (https://github.com/bxparks/AceCRC).
+ * DO NOT EDIT
+ */
+#include "crc8_bit.hpp" // header file converted by AceCRC
+#include <stdlib.h>
+#include <stdint.h>
+#include <stdbool.h>
+
+namespace ace_crc {
+namespace crc8_bit {
+
+
+
+crc_t crc_update(crc_t crc, const void *data, size_t data_len)
+{
+    const unsigned char *d = (const unsigned char *)data;
+    unsigned int i;
+    bool bit;
+    unsigned char c;
+
+    while (data_len--) {
+        c = *d++;
+        for (i = 0x80; i > 0; i >>= 1) {
+            bit = crc & 0x80;
+            if (c & i) {
+                bit = !bit;
+            }
+            crc <<= 1;
+            if (bit) {
+                crc ^= 0x07;
+            }
+        }
+        crc &= 0xff;
+    }
+    return crc & 0xff;
+}
+
+} // crc8_bit
+} // ace_crc